+++ src/main/java/com/mageddo/dnsproxyserver/config/entrypoint/ConfigJson.java
@@ -38,4 +38,6 @@ public interface ConfigJson {
   SimpleServer.Protocol getServerProtocol();
 
   URI getDockerHost();
+
+  Boolean getResolvConfOverrideNameServers();
 }