+++ RELEASE-NOTES.md
@@ -9,6 +9,7 @@ for more details
 * Cache improvement for a concurrent approach synchronized by key
 * Increasing tcp and udp server parallelism from 20 to 50 threads each, threads are created lazilly
 * Fixing solver NPE on unsupported query types
+* Option to not comment out existing nameservers from resolv.conf [see the docs](http://mageddo.github.io/dns-proxy-server/3.15/en/3-configuration/)
 
 ## 3.14.5
 * Specify minimum required docker api version `1.21` (as DPS 2)